python_api/requirements.txt
fastapi
uvicorn
python-dotenv
google-generative-ai
yt-dlp
instaloader
requests
aiohttp
aiofiles
python-multipart
